insert board ..ing..

+++ server/src/main/java/com/ssafy/groupbuying/repository/BoardRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.groupbuying.vo.Board;
 
-public interface BoardRepository extends JpaRepository<Board, Integer> {
-	public Board findById(int id);
+public interface BoardRepository extends JpaRepository<Board, Long> {
+	public Board findById(long id);
 
 }
